Update documentation and Makefile to support Markdown output
Added a new target for generating Markdown files in the Makefile and updated the help message accordingly. Also, included the sphinx-markdown-builder dependency in requirements.txt. Changed instances of ".. tip::" to ".. note::" in parsing.rst and writing.py for consistency in documentation style. Updated regex patterns in filters.py and intraword.py to escape backslashes correctly.